# openapi_schema.yaml
openapi: 3.0.3
info:
  title: API ERP El Arbolito (Facturación SRI)
  version: 1.0.0
  description: API para gestión de socios, lecturas de agua y facturación electrónica
    con firma XAdES-BES.
paths:
  /api/v1/analytics/cartera-vencida/:
    get:
      operationId: api_v1_analytics_cartera_vencida_retrieve
      description: Retorna la lista de socios con deuda, clasificada por antigüedad
        (Corriente, 1-3 meses, >3 meses).
      summary: Reporte de Cartera Vencida (Aging Report)
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/analytics/cierre-caja/:
    get:
      operationId: api_v1_analytics_cierre_caja_retrieve
      description: Retorna el total recaudado hoy (o en rango de fechas), desglosado
        por Efectivo/Transferencia.
      summary: Cierre de Caja Diario
      parameters:
      - in: query
        name: fecha_fin
        schema:
          type: string
          format: date
        description: YYYY-MM-DD
      - in: query
        name: fecha_inicio
        schema:
          type: string
          format: date
        description: YYYY-MM-DD
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/v1/analytics/dashboard-kpis/:
    get:
      operationId: api_v1_analytics_dashboard_kpis_retrieve
      description: Métricas rápidas para la pantalla de inicio del Tesorero.
      summary: Dashboard KPIs (Ejecutivo)
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT